+++ docs/versionhistory.rst
@@ -6,6 +6,8 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 **UNRELEASED**
 
 - Added the ``start_background_task()`` function
+- Allowed retrieving existing resources from a context in teardown callbacks (this would
+  previously raise ``RuntimeError("this context has already been closed")``)
 
 **4.8.0** (2022-04-28)